Added routes for Profile
Added CRUD routes  for the profile resource. A GET route to get profile of logged in user and a POST route to create/update a user profile. Added routes to get all users, to get user via id, to create a user profile and to delete a user profile.Added routes to create a profile experience and to delete a profile experience